requirements.txt
opencv-python==4.5.3.56
torch==1.12.1
torchvision==0.13.1
pillow==8.2.0
numpy==1.19.5
pandas==1.3.0
scipy==1.5.4